STYLE: Fix build warnings on linux
Fixed build warnings reported by this github action:
https://github.com/Slicer/Slicer/actions/runs/11829892124/job/32962501349

+++ Libs/MRML/Core/vtkMRMLClipNode.cxx
@@ -556,7 +556,7 @@ int vtkMRMLClipNode::GetClippingNodeState(const char* nodeID)
 int vtkMRMLClipNode::GetNthClippingNodeState(int n)
 {
   NodeReferenceListType references = this->NodeReferences[ClippingNodeReferenceRole];
-  if (references.size() <= n)
+  if (references.size() <= static_cast<size_t>(n))
   {
     return ClipOff;
   }
@@ -596,7 +596,7 @@ void vtkMRMLClipNode::SetClippingNodeState(const char* nodeID, int state)
 void vtkMRMLClipNode::SetNthClippingNodeState(int n, int state)
 {
   NodeReferenceListType references = this->NodeReferences[ClippingNodeReferenceRole];
-  if (references.size() <= n)
+  if (references.size() <= static_cast<size_t>(n))
   {
     return;
   }
